% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{as.data.frame.InferenceResult}
\alias{as.data.frame.InferenceResult}
\title{Convert InferenceResult to data.frame}
\usage{
\method{as.data.frame}{InferenceResult}(x, ...)
}
\arguments{
\item{x}{An InferenceResult object}
\item{...}{Additional arguments (ignored)}
}
\value{
A data.frame with samples
}
\description{
Convert InferenceResult to data.frame
}
