% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Bernoulli}
\alias{Bernoulli}
\title{Bernoulli Distribution}
\usage{
Bernoulli(p)
}
\arguments{
\item{p}{Probability of success (can be a number or parameter name)}
}
\value{
A distribution specification
}
\description{
Create a Bernoulli distribution for binary outcomes (0 or 1).
}
\examples{
Bernoulli(0.5)       # 50\% success rate
Bernoulli("theta")   # Parameter from model
}
