% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Beta}
\alias{Beta}
\title{Beta Distribution}
\usage{
Beta(alpha, beta)
}
\arguments{
\item{alpha}{Shape parameter alpha (concentration1)}
\item{beta}{Shape parameter beta (concentration0)}
}
\value{
A distribution specification
}
\description{
Create a Beta distribution. Defined on (0, 1), useful for probability parameters.
}
\examples{
Beta(1, 1)   # Uniform on [0, 1]
Beta(2, 5)   # Skewed toward 0
}
