% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{BetaBinomial}
\alias{BetaBinomial}
\title{Beta-Binomial Distribution}
\usage{
BetaBinomial(n, alpha, beta)
}
\arguments{
\item{n}{Number of trials (positive integer)}
\item{alpha}{First shape parameter of the Beta prior (alpha > 0)}
\item{beta}{Second shape parameter of the Beta prior (beta > 0)}
}
\value{
A bg_distribution object
}
\description{
Create a Beta-Binomial distribution. A compound distribution where
the success probability follows a Beta distribution. Useful for
overdispersed binomial data.
}
\examples{
\dontrun{
BetaBinomial(10, 2, 5)  # n=10, alpha=2, beta=5
}
}
