% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Binomial}
\alias{Binomial}
\title{Binomial Distribution}
\usage{
Binomial(n, p)
}
\arguments{
\item{n}{Number of trials}
\item{p}{Probability of success (can be a number or parameter name)}
}
\value{
A distribution specification
}
\description{
Create a Binomial distribution for count of successes in n trials.
}
\examples{
\donttest{
Binomial(100, 0.5)      # 100 trials, 50\% success
Binomial(100, "theta")  # Parameter from model
}
}
