% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Exponential}
\alias{Exponential}
\title{Exponential Distribution}
\usage{
Exponential(rate)
}
\arguments{
\item{rate}{Rate parameter (inverse of mean)}
}
\value{
A distribution specification
}
\description{
Create an Exponential distribution. For waiting times and durations.
}
\examples{
\dontrun{
Exponential(1)  # Rate of 1
}
}
