% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{LogNormal}
\alias{LogNormal}
\title{Log-Normal Distribution}
\usage{
LogNormal(loc, scale)
}
\arguments{
\item{loc}{Mean of the log (mu)}
\item{scale}{Standard deviation of the log (sigma)}
}
\value{
A distribution specification
}
\description{
Create a Log-Normal distribution. For positive values with multiplicative effects.
}
\examples{
\donttest{
LogNormal(0, 1)
}
}
