% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Normal}
\alias{Normal}
\title{Normal Distribution}
\usage{
Normal(loc, scale)
}
\arguments{
\item{loc}{Mean (can be a number or parameter name as string)}
\item{scale}{Standard deviation (can be a number or parameter name)}
}
\value{
A distribution specification (JSON string internally)
}
\description{
Create a Normal (Gaussian) distribution specification.
}
\examples{
Normal(0, 1)           # Standard normal
Normal("mu", "sigma")  # Parameters from model
}
