% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Poisson}
\alias{Poisson}
\title{Poisson Distribution}
\usage{
Poisson(rate)
}
\arguments{
\item{rate}{Rate parameter (expected count, can be a number or parameter name)}
}
\value{
A distribution specification
}
\description{
Create a Poisson distribution for count data.
}
\examples{
\dontrun{
Poisson(5)          # Rate of 5
Poisson("lambda")   # Parameter from model
}
}
