% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{print.bg_distribution}
\alias{print.bg_distribution}
\title{Print a BayesianGPU Distribution}
\usage{
\method{print}{bg_distribution}(x, ...)
}
\arguments{
\item{x}{A bg_distribution object}
\item{...}{Additional arguments (ignored)}
}
\description{
Print method for bg_distribution objects.
}
