% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{summary.InferenceResult}
\alias{summary.InferenceResult}
\title{Summary of InferenceResult}
\usage{
\method{summary}{InferenceResult}(object, ...)
}
\arguments{
\item{object}{An InferenceResult object}
\item{...}{Additional arguments (ignored)}
}
\value{
A data.frame with summary statistics
}
\description{
Summary of InferenceResult
}
