% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{TruncatedNormal}
\alias{TruncatedNormal}
\title{Truncated Normal Distribution}
\usage{
TruncatedNormal(loc, scale, low, high)
}
\arguments{
\item{loc}{Location parameter (mean of underlying normal)}
\item{scale}{Scale parameter (std dev of underlying normal, must be positive)}
\item{low}{Lower bound}
\item{high}{Upper bound}
}
\value{
A distribution specification
}
\description{
Create a Truncated Normal distribution. A Normal distribution
constrained to the interval [low, high].
}
\examples{
\donttest{
TruncatedNormal(0, 1, -2, 2)  # Normal(0,1) truncated to [-2, 2]
}
}
