% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Uniform}
\alias{Uniform}
\title{Uniform Distribution}
\usage{
Uniform(low, high)
}
\arguments{
\item{low}{Lower bound}
\item{high}{Upper bound}
}
\value{
A distribution specification
}
\description{
Create a Uniform distribution.
}
\examples{
\dontrun{
Uniform(0, 10)  # Uniform on [0, 10]
}
}
